backend/requirements/base.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
cryptography>=42.0.0
sqlalchemy==2.0.23
aiosqlite==0.19.0
structlog==23.2.0
typer==0.9.0
rich==13.7.0
google-generativeai==0.3.2
python-multipart==0.0.6
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
watchdog==3.0.0
python-dotenv==1.0.0